Screenshots at maximum zoom, and the two defects that only a screenshot could find
The operator, 2026-08-22: "Can you confirm that rendering on screen is actually
happening at maximum zoom? zoom in on one of the michocondria structures and
post screenshots here to confirm."

It was not. Two defects, and nothing in the suite could see either.

WHY NOTHING ALREADY THERE COULD ANSWER IT
------------------------------------------
`zooming_does_not_throw_away_where_the_operator_panned` proves the view stays
where it is put to a trillion percent. `zooming_past_the_pixmap_ceiling_still_
renders` proves no raster is refused. NEITHER LOOKS AT THE SCREEN, and a canvas
can satisfy both while drawing blank paper: the arithmetic would be perfect, the
rasters would complete, and the operator would see nothing. That is exactly what
was happening past about 10^7 %.

O24i -- THE REGION PATH NARROWED TO f32
----------------------------------------
`render::strategy::region_for` snaps the region origin to a half-view grid:

    snapped_x = (x0 / step_x).floor() * step_x

At a trillion percent `step_x` is about 2e-8 pt while `x0` is an ordinary page
coordinate near 540. Their quotient is 2e10 -- past f32's last exactly
representable integer of 2^24 -- so `.floor()` was applied to a number that had
already lost its integer part.

Measured before the fix: from about 10^7 % the region stopped shrinking and
floored at 2.4414e-3 x 3.0213e-3 pt, FIFTY THOUSAND TIMES the 4.8e-8 x 6.2e-8
the viewport was showing, and its raster was painted 18,998,834 window points
off the viewport. `drawn=1` was still traced and no render failed, so every
check passed while the operator saw a fraction of one texel stretched across the
window.

The whole path is f64 now -- region_for, overscanned, page_region, OVERSCAN, and
the canvas call site that cast `DeepAnchor::visible_rect`'s f64 result straight
down to f32. That cast was the one narrowing left in a path whose every other
stage was already f64, and it was narrowing the value the tier exists to
compute.

The real ceiling of the fixed design, since it is worth knowing: the extent is
`(x0 + w) - x0` at an absolute position near 540, where an f64 ULP is 1.1e-13.
At maximum zoom `w` is 1e-9 pt -- about 8,800 ULPs, eighteen representable steps
per screen pixel. The tier below ran out at 2^24; this one has room left.

O24j -- THE STATUS BAR SHOWED 4294967295%
------------------------------------------
`ViewState::zoom_percent` returned a u32 and `as u32` saturates, so past about
42,949,672 % the readout showed u32::MAX presented as a measurement. Seen in the
screenshot gallery, which is the only instrument that reads the number an
operator reads.

The type was right when MAX_ZOOM was 8.0 and every reachable value fitted in
three digits. O24 raised the ceiling to 10^12 and did not revisit it -- the
recurring shape of this whole request: a limit lifted in one place while a
narrower type downstream keeps enforcing the old one silently.

It reports 999999995904% at the top, not 1000000000000%, and that is correct
rather than a rounding failure: ViewState::zoom is an f32, so that is the
nearest representable value and it is what the view actually is. Pinned exactly,
so a change that starts rounding the display instead of reporting it has to be
deliberate.

THE GALLERY
-----------
`the_page_still_renders_at_every_decade_of_zoom` -- new. Parks the pointer on a
DOCUMENT coordinate, climbs by Ctrl+wheel and photographs the window at each
tier of the fixture's own scale chain. Three assertions per step, because any
two can hold while the third fails:

  the CANVAS REGION is not near-uniform   rules out a blank page
  the canvas traced drawn >= 1            rules out space reserved with no raster
  no outcome=failed render                rules out a swallowed refusal

The canvas region, NOT the window. The first version asked
`capture::window_to_png`, which refuses a near-uniform WINDOW -- and a window
always contains a ribbon and two panels, so it can never fire for the reason
this check cares about. It passed a screenshot of blank white paper on that
technicality. Third instance this session of an assertion aimed at the wrong
surface.

It also RE-AIMS between tiers, from the f64 position line. Zoom-to-cursor holds
the point to about half a per-notch tolerance, which is excellent per notch and
still accumulates over the ~120 notches to the ceiling -- so without it the run
wanders off a 3 um mitochondrion and photographs cytoplasm. CanvasMapping cannot
do the re-aiming: it converts through the f32 `rect=`, whose representable
spacing at the ceiling is half a million points.

WHAT THE SCREENSHOTS SHOW
-------------------------
        114 %  the whole banana                                    318 tones
      2,785 %  the two cells                                       293
     13,794 %  cell labels, organelles                             487
     45,799 %  labelled organelles                                 707
    504,845 %  one mitochondrion in cytoplasm, cell wall behind    144
  3,730,330 %  mitochondria with cristae                           238
 41,120,084 %  mtDNA nucleoid, mitoribosomes, ATP synthase heads    222
999,999,995,904 %  mitochondrial matrix, a 0.02 nm field            15

The last row is a solid fill and that is CORRECT: at the ceiling the viewport
spans 6e-8 pt, smaller than an atom, and the fixture has nothing below the 10 nm
ATP synthase heads. Rendering is still happening; there is nothing left to draw.

Gates 17/17. Tests 1,688. Two new unit tests pin the region's proportionality
AND its origin -- the size test alone would pass on an implementation that
returned a correctly-sized rect somewhere else entirely, which is close to what
the f32 version did.

+++ crates/pdfce-gui/src/render/region.rs
@@ -52,17 +52,15 @@ use pdfce_core::page_tree::Rect;
 /// why that is the difference between panning smoothly and waiting for a redraw
 /// on every pixel of movement.
 #[must_use]
-pub fn page_region(visible_canvas: (f32, f32, f32, f32), page_pts: (f32, f32)) -> Rect {
+/// ★★ `visible_canvas` is `f64`: at deep zoom it holds a rectangle a few times
+/// 10⁻⁸ pt wide at an absolute position near 540, and `f32` cannot carry both
+/// magnitudes at once. See [`super::strategy::region_for`].
+pub fn page_region(visible_canvas: (f64, f64, f64, f64), page_pts: (f32, f32)) -> Rect {
     let (x0, y0, x1, y1) = super::strategy::region_for(visible_canvas);
     let height = f64::from(page_pts.1);
     // y-down from the top becomes y-up from the bottom, so the two y values
     // also swap ends: the canvas's smaller y is the PDF's larger one.
-    Rect::from_corners(
-        f64::from(x0),
-        height - f64::from(y1),
-        f64::from(x1),
-        height - f64::from(y0),
-    )
+    Rect::from_corners(x0, height - y1, x1, height - y0)
 }
 
 /// Where a region's raster belongs on screen, given where the whole page would
@@ -183,7 +181,7 @@ mod tests {
             (back.3, wanted.3),
         ] {
             assert!(
-                (got - want).abs() < 0.01,
+                (f64::from(got) - want).abs() < 0.01,
                 "round trip lost the rect: {back:?} vs {wanted:?}"
             );
         }
